Add test to see if PubSub item can be deleted (#5)
* Add test to see if PubSub item can be deleted

XEP-0060 defines optional functionality to delete items from a node. This commit adds a test for this, that's only executed when the corresponding feature is announced through service discovery.

As an aside, this tests the fix for Openfire's OF-2791 issue.

* Add human-readable message to assertion in new PubSub test

This message intends to help people to determine the exact cause of the test failure, when they have access to the XMPP log, without having to look at the test implementation.