Add ability to store XMPP logs in files
This adds a new Smack debug logger that can be used to store XMPP traffic in distinct files per test.

To configure the logger, use these system properties when invoking the tests:

```
-Dsinttest.debugger="org.igniterealtime.smack.inttest.util.FileLoggerFactory"
-DlogDir=target/logs
```

The `logDir` property value is used to identify the directory in which the logs will be stored. The directory will be created if it doesn't already exist.