De-duplicate PubSubIntegrationTest
Both this project and the Smack project define the class `org.jivesoftware.smackx.pubsub.PubSubIntegrationTest`. It has been observed that as a result, re-packaging the project can overwrite one class with the other.

To prevent that issue, the class in this project has been renamed. Duplicate tests have been removed.