Improve test run output on standard-out
The output as printed on standard-out should facilitate viewers in debugging their code on test failures.

It needs to be as easy as possible to determine for each test failure:
- which specification was the subject of the test, containing a specification reference and a direct quote of the relevant text.
- what test assertion failed
- where the XMPP stanza log is for the test that fails.

Viewers need not be familiar with our test implementations (or even Smack, or Java). We can/should still reference the implementation as a last resort.

This is achieved by replacing SINT's standard Test Run Result processor (`org.igniterealtime.smack.inttest.SmackIntegrationTestFramework\$JulTestRunResultProcessor`) with a new, custom implementation: `org.igniterealtime.smack.inttest.util.StdOutTestRunResultProcessor`