Add tests for XEP-0421: Ensure that stanzas contain no more than 1 occupant-id
In this commit, the preexisting tests for XEP-0421 are duplicated, but with a different assertion. The new tests now assert that no more than one occupant-id element is present (as having more than one is a MUST NOT according to section 4 of the XEP).

To avoid code duplication, this commit introduces 're-usable' fixtures. These fixtures perform all setup and teardown, and are provided an assertion. The tests invoke the fixture with the assertion. That allows multiple tests to use the same test fixture (but asserting different things), while keeping code duplication low.