XEP-0133: Improve ad-hoc command execution status assertion
XEP-0050 and XEP-0133 are not overly clear in their representation of a successful command execution. Generally, a command can fail in two ways:
1. On a stanza level (eg: IQ result=error)
2. With a 'note' on the command status of type error

Not all server implementations can be trusted to use the command status note to signal failure. Ejabberd, for example, doesn't use those. On the other hand, the IQ type should always be checked anyway.

This commit:
- removes Openfire-specific checks for success or failures (that depends on specific wording of notes)
- introduces re-usable assertions to validate succes or failure of an ad-hoc command completion
- replace some variations of command completion outcome assertion with one of these two assertions
- add command completion assertions to some tests that didn't have them yet