XEP-0133: Remove user in test fixture cleanup
Whenever user creation is attempted in a test (either in the test fixture setup, or as part of executing the system under test), the test teardown should attempt to remove that user. This helps keeping the test environment clean, which reduces the chance of tests affecting each-other.

The test should not error when the user cannot be removed (as the user may not have been created).