Add tests for RFC 6121 Section 8.5.2.1.1 'Server Rules for Processing XML Stanzas'
Integration tests that verify that behavior defined in section 8.5.2.1.1 "Local User / localpart@domainpart / Available or Connected Resources / Message" of section 8 "Server Rules for Processing XML Stanzas" of RFC6121.

This commit refactors the previous commit for code maintainability. It replaces all duplicated code with one method that performs the execution of each test, including test fixture setup and teardown. The assertions are passed to this method as an argument.