Upgrade to Smack-4.5.0-beta6: Use try/close for StanzaListener deregistration
When registering a StanzaListener with a connection, that instance should be deregistered again after use. This helps to prevent issues where the execution of one test affects another test (tests, after all, can share the same connection).

In Smack-4.5.0-beta6, the API for StanzaListener (de)registration was modified. It can now use Java's try-with-resources mechanism. This helps to ensure that each StanzaListener is deregistered after use.

In this commit, all StanzaListener usages have been modified to use this new mechanism. This includes several instances of listeners that were not deregistered properly before.

Local testing suggests that this commit reduces the amount of occassional, hard-to-reproduce test failures.