Fix #110: Inconsistent results for RosterIntegrationTest
The RosterIntegrationTest frequently fails. Through debugging, we've established that this happens when the Roster's IQRequestHandler that processes roster pushes is not invoked (even though a corresponding stanza is received).

The source for this problem has been tracked down to RFC6121Section8_5_3_2_3_IqIntegrationTest. This test replaces IQRequestHandlers with an instance that's used by the test. The intention is to restore the handler that was replaced after the test has finished execution.

This commit fixes a bug that prevented the restoration of the original IQRequestHandler.

In the test, the handlers for multiple connections are replaced. The old values, however, are not stored _per connection_. This causes some connections to not get their original handler restored.