Skip ad-hoc tests when not provisioned with an admin account
If the runner is configured to use three explicitly named accounts (instead of an admin account that creates accounts as desired), the tests for ad-hoc commands crash. These tests require an administrative account.

In this comment, these tests raise a TestNotPossibleException, rather than cause a crash. This allows other tests to continue.