Add config options to provision test accounts
Instead of using an admin account or in-band registration to provision test accounts, the runner can be configured with three distinct test accounts.

In this commit, new input parameters are added to the entrypoint script for the containers that can be used by an end-user to configure those accounts.

fixes #126