No longer check for `jabber:iq:roster` disco#info feature
The XSF Registrar registers jabber:iq:roster as a service discovery feature in https://xmpp.org/registrar/disco-features.html but none of the relevant RFCs mandate that this feature is to be advertised.

RFC 6121 section 12 suggests that support of roster (get/set) is mandatory for servers.

Some of our tests check for this feature to be advertised, and throw a TestNotPossible if that's not the case. This check should be removed.

fixes #144