Don't expect to be able to crawl the entire Service Discovery hierarchy
ServiceDiscoveryIntegrationTest has a couple of tests that verify aspects of the hierarchy of service discovery information. This test throws a TestNotPossible whenever any node in the hierarchy can't be queried.

It is probably reasonable for a server to start throwing 'not-allowed' errors when crawling certain parts on the hierarchy. That should not necessarily be a reason to skip this particular test. Maybe 'enough' of the hierarchy was queried for the test to be able to execute.

fixes #147