Revert "Disable Smack-provided 'client' tests"
This reverts commit db2312a676d41ad2af53cc23655f66171fcb17fc.

Although the reverted solution will probably work, it needs to be kept consistently configured in all runners. I'm also not sure if this affects the flexibility with the various enable and disable options that we present to the user.

This solution is being replaced by one that is based on generating a shaded jar, that will exclude the offending tests.