Exclude Smack-provided 'client' tests
Our integration test suite started with a number of tests that were developed for the Smack project. The original purpose of these tests was to validate Smack's operation, integrated with various different servers.

Some of these tests aren't very appropriate to our server integration test suite, as they test client-interaction. Such tests should be excluded from our test set.

This approach depends on the Maven 'shade' plugin, which generates an 'uber jar', that potentially is modified. This commit applies such a modification: filter out the tests that we don't want.

The resulting 'uber jar' is very similar to the 'jar-with-dependencies' that we already generated. The old configuration that produced the 'jar-with-dependencies' has been removed, in favor of the 'uber-jar' that's generated by the shader plugin. That jar has been named 'jar-with-dependencies' so that our runners will see the same file name.

One caveat of changing the dependency-provided jar this way, is that their signatures no longer match with the content. To prevent security exceptions, the shader is configured to remove the affected signatures.